requirements.txt
flask
pymongo
nltk
requests
python-dotenv
isodate 